✨ Lots of reworking and a new RandomSecret component.
The core framework is now in core/ and the components are in components/. This avoids an import loop between core and test because I want to use the test tools for the components.

This does then create a (smaller) import loop between core and components which is solved with some minor dependency injection at startup. If go supports Python-style late binding imports it would solve this but oh well, this is good enough.