✨ A lot of work.
Adding a backport of k/k 1.19's metav1.Condition struct so I can use it now and change it to an alias later. And then add in some real status condition support in both the template component and a new ready status component. Also solves a bunch of test flake issues with the random secret component.